requirements.txt
mysql-connector-python
ollama
chromadb
python-dotenv
